+++ web/src/Pages/Public/Lectures/Lectures.styles.jsx
@@ -1,0 +1,9 @@
+import {makeStyles} from '@material-ui/styles';
+
+export const useStyles = makeStyles((theme) => ({
+  root: {},
+  paper: {
+    height: 700,
+    padding: theme.spacing(1),
+  },
+}));
