+++ web/src/navconfig.jsx
@@ -26,7 +26,7 @@ import Submissions from './Pages/Public/Submissions';
 import Submission from './Pages/Public/Submission';
 import Blog from './Pages/Public/Blog';
 import Visuals from './Pages/Public/Visuals/Visuals';
-import Lectures from './Pages/Public/Lectures';
+import Lectures from './Pages/Public/Lectures/Lectures';
 
 import AdminUsers from './Pages/Admin/Users';
 import AdminUser from './Pages/Admin/User';
